requirements.txt
absl-py==1.3.0
asttokens==2.0.8
astunparse==1.6.3
backcall==0.2.0
cachetools==5.2.0
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==2.1.1
colorama==0.4.5
contourpy==1.0.5
customized-konlpy==0.0.64
cycler==0.11.0
dataclasses==0.6
debugpy==1.6.3
decorator==5.1.1
dicom2jpg==0.1.10
distlib==0.3.6
docopt==0.6.2
entrypoints==0.4
executing==1.1.0
filelock==3.8.0
flatbuffers==22.9.24
fonttools==4.38.0
fuzzywuzzy==0.18.0
gast==0.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.3.post0
grpcio==1.50.0
h5py==3.7.0
huggingface-hub==0.10.1
idna==3.4
imageio==2.22.4
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
JPype1==1.4.0
jupyter_client==7.3.5
jupyter-core==4.11.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
konlpy==0.6.0
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mritopng==2.2
mysql-connector-python==8.0.30
nest-asyncio==1.5.6
networkx==3.1
nibabel==4.0.2
numpngw==0.1.2
numpy==1.23.2
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pip==22.0.4
pipdeptree==2.7.1
pipreqs==0.4.13
platformdirs==2.6.0
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydicom==2.3.1
Pygments==2.13.0
pylibjpeg==1.4.0
pylibjpeg-libjpeg==1.3.2
pylibjpeg-openjpeg==1.3.0
pymongo==4.1.1
PyMySQL==1.0.2
pyparsing==3.0.9
pypng==0.20220715.0
PyTDC==0.4.0
python-dateutil==2.8.2
pytz==2022.2.1
pywin32==304
PyYAML==6.0
pyzmq==24.0.1
rdkit-pypi==2022.9.5
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.1
setuptools==58.1.0
six==1.16.0
sklearn==0.0
split-folders==0.5.1
SQLAlchemy==1.4.42
stack-data==0.5.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.10.0
tensorflow-addons==0.18.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tf==1.0.0
threadpoolctl==3.1.0
timm==0.5.4
tokenizers==0.13.1
torch==1.13.1
torch-geometric==2.3.1
torch-tb-profiler==0.4.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
transformers==4.23.1
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.12
virtualenv==20.17.1
wcwidth==0.2.5
Werkzeug==2.2.2
wheel==0.37.1
wrapt==1.14.1
yarg==0.1.9
